fix previous session not being resumed when tab is reloaded (#40)
## Summary

Running `@launchdarkly/observability` and `@launchdarkly/session-replay`
plugins
concurrently would cause the session id to be cleared by one before the
other plugin
would be able to read it. This PR fixes the behavior to make sure that
the
session is shared between the two plugins.

Loading the `@launchdarkly/session-replay` plugin and reloading the page
before
the session would initialize would start a new session on the next
reload as
the stored session ID would be deleted. This PR also corrects the
behavior.

## How did you test this change?


https://www.loom.com/share/c9de3d20c0bf4cfb85910170306241a6?from_recorder=1&focus_title=1

## Are there any deployment considerations?

changeset 

## Does this work require review from our design team?

no