fix JS GHA not auto-publishing LD packages (#49)
## Summary

`turbo.yml` GHA currently only auto-publishes the highlight SDKs, not
the launchdarkly ones.
This is a problem since they are interdependent and should be published
together if there are new versions released.

`yarn changeset version` is unrelated to npm publishing (creates PR that
will bump versions if changesets are detected)
but should happen first / separately for clarity.

## How did you test this change?

CI

## Are there any deployment considerations?

no

## Does this work require review from our design team?

no