feat: Add basic config for dotnet plugin. (#167)
## Summary

This PR adds a basic config for the dotnet plugin and an initial otel
setup.

This otel setup does not include being packaged as a plugin or the
export sampling configuration.

It also includes an ASP.Net application to use for development.
Currently this is in the same solution for ease of development. It could
move to the e2e projects if needed later.

## How did you test this change?

<!--
Frontend - Leave a screencast or a screenshot to visually describe the
changes.
-->

## Are there any deployment considerations?

<!--
 Backend - Do we need to consider migrations or backfilling data?
-->