feat: Add graphql query to get sampling config. (#181)
## Summary

Uses the sampling tace and log exporter with the sampling configuration.

The approach used for GraphQL is a bit different on this one. The types
are not tool generated, so they would require updating, but it does
remove any runtime dependency. I may consider back-porting this approach
to reduce dependencies in the other plugins.

## How did you test this change?

Unit tests.
Manual testing.

## Are there any deployment considerations?

<!--
 Backend - Do we need to consider migrations or backfilling data?
-->