feat: O11Y-359 - Add custom sampling for OTLP logs and traces (#189)
## Summary

This commit introduces a custom sampling mechanism for OpenTelemetry
(OTLP) logs and traces within the Android observability SDK. This allows
for more granular control over which telemetry data is exported, helping
to manage data volume and focus on relevant information.

The architecture and implementation are mostly based on the
[observability-node](https://github.com/launchdarkly/observability-sdk/tree/main/sdk/%40launchdarkly/observability-node)
project

The key additions include:
- `CustomSampler`: A class that implements sampling logic based on a
`SamplingConfig`.
- `SamplingConfig`: Data classes defining the structure for configuring
sampling rules for spans and logs.
- `ExportSampler`: An interface defining the contract for samplers that
operate during the export phase.
- `SamplingTraceExporter` and `SamplingLogExporter`: Decorator classes
that wrap the standard `OtlpHttpSpanExporter` and
`OtlpHttpLogRecordExporter` respectively. They use the `CustomSampler`
to filter spans/logs before delegating to the underlying exporter.
- `SamplingResult`: A data class to hold the outcome of a sampling
decision (whether to sample and any additional attributes).
- `sampleSpans` and `sampleLogs`: Utility functions that take a list of
spans or logs and an `ExportSampler` to filter them based on sampling
decisions.

What is not included:
- Fetching sampling configuration from the backend service
- Setting sampling configuration to the CustomSampler

## How did you test this change?
Unit testing

## Are there any deployment considerations?
Since the sample configuration isn't being retrieved from the backend
yet, null is currently passed as the config. This means no data will be
sampled.

---------

Co-authored-by: Todd Anderson <127344469+tanderson-ld@users.noreply.github.com>