fix: Fix an issue where metrics could have NaN or infinite values. (#220)
## Summary

Fix an issue where the `downlinkMax` metric could be infinite and
prevent successful metric exports.

Infinite/NaN would become `null` in JSON, and the collector will treat
the entire request payload as invalid if there is a null metric payload.

## How did you test this change?

<!--
Frontend - Leave a screencast or a screenshot to visually describe the
changes.
-->

## Are there any deployment considerations?

<!--
 Backend - Do we need to consider migrations or backfilling data?
-->