feat: Support Go observability initialization independent of plugins. (#230)
## Summary

Allow for initialization of Go observability independent of the
LaunchDarkly SDK.

This allows for advanced use-cases where the LD SDK may not be in use,
or may not be readily accessible.

I called it `PreInitialize`, but I am not attached to that name. 

You do this:
```go
	// Initialize the observability plugin ahead of the LaunchDarkly client.
	// This is generally only required for advanced use cases and is not the
	// recommended way to initialize observability.
	// Observability can also be used without the LaunchDarkly client at all
	// by using this method. Some features will not be available when using
	// observability without the LaunchDarkly client.
	ldobserve.PreInitialize(os.Getenv("LAUNCHDARKLY_SDK_KEY"),
		ldobserve.WithEnvironment("test"),
		ldobserve.WithServiceName("go-plugin-example"),
		ldobserve.WithServiceVersion(version.Commit),
	)
```

And now observability it setup. 

You can then later optionally use the plugin to facilitate registering
hooks.

```
	client, _ := ld.MakeCustomClient(os.Getenv("LAUNCHDARKLY_SDK_KEY"),
		ld.Config{
			Plugins: []ldplugins.Plugin{
				// This special form of constructing the observability plugin
				// is used when observability has been pre-initialized.
				ldobserve.NewObservabilityPluginWithoutInit(),
			},
		}, 5*time.Second)
```

Or you can directly use the ldotel hook instead. Using the plugin will
be more resilient in case we add more hooks.


## How did you test this change?

Manual testing.