feat: Add basic flutter plugin. (#233)
## Summary

Adds a basic observability plugin for flutter. This doesn't have any
configuration and only support basic tracing.

There are wrappers for spans, attributes, and the span kind because
there isn't an official open telemetry implementation for dart. We don't
want to encourage importing the API package as is supported in most
other SDKs.

We may want to consider vendoring the otel code and keeping it internal
to limit the potential breakage if we change implementations.

## How did you test this change?

Manual testing. Unit tests.

## Are there any deployment considerations?

<!--
 Backend - Do we need to consider migrations or backfilling data?
-->